requirements.txt
transformers
streamlit
torch
python-dotenv